Checkpoint: moved constants.py under core/
