+++ manage_arkime.py
@@ -11,7 +11,7 @@ from commands.destroy_demo_traffic import cmd_destroy_demo_traffic
 from commands.get_login_details import cmd_get_login_details
 from commands.list_clusters import cmd_list_clusters
 from commands.remove_vpc import cmd_remove_vpc
-import constants as constants
+import core.constants as constants
 from core.capacity_planning import MAX_TRAFFIC, DEFAULT_SPI_DAYS, DEFAULT_REPLICAS, DEFAULT_S3_STORAGE_DAYS, DEFAULT_HISTORY_DAYS
 from logging_wrangler import LoggingWrangler, set_boto_log_level
