+++ manage_arkime/aws_interactions/s3_interactions.py
@@ -4,7 +4,7 @@ import logging
 from botocore.exceptions import ClientError
 
 from aws_interactions.aws_client_provider import AwsClientProvider
-import constants as constants
+import core.constants as constants
 
 logger = logging.getLogger(__name__)
