+++ manage_arkime/commands/add_vpc.py
@@ -5,7 +5,7 @@ import aws_interactions.ec2_interactions as ec2i
 import aws_interactions.events_interactions as events
 import aws_interactions.ssm_operations as ssm_ops
 from cdk_interactions.cdk_client import CdkClient
-import constants as constants
+import core.constants as constants
 import cdk_interactions.cdk_context as context
 from vni_provider import SsmVniProvider, VniAlreadyUsed, VniOutsideRange, VniPoolExhausted
