+++ manage_arkime/vni_provider.py
@@ -3,7 +3,7 @@ import json
 import logging
 from typing import Dict, List
 
-import constants as constants
+import core.constants as constants
 from aws_interactions.aws_client_provider import AwsClientProvider
 import aws_interactions.ssm_operations as ssm_ops
