+++ test_manage_arkime/commands/test_add_vpc.py
@@ -7,7 +7,7 @@ from aws_interactions.aws_environment import AwsEnvironment
 import aws_interactions.ec2_interactions as ec2i
 import aws_interactions.events_interactions as events
 from aws_interactions.ssm_operations import ParamDoesNotExist
-import constants as constants
+import core.constants as constants
 import vni_provider as vnis
