+++ test_manage_arkime/commands/test_destroy_demo_traffic.py
@@ -2,7 +2,7 @@ import unittest.mock as mock
 
 from aws_interactions.aws_environment import AwsEnvironment
 from commands.destroy_demo_traffic import cmd_destroy_demo_traffic
-import constants as constants
+import core.constants as constants
 
 @mock.patch("commands.destroy_demo_traffic.AwsClientProvider")
 @mock.patch("commands.destroy_demo_traffic.CdkClient")
