+++ test_manage_arkime/lambda_destroy_eni_mirror/test_destroy_eni_mirror_handler.py
@@ -4,7 +4,7 @@ import unittest.mock as mock
 from lambda_destroy_eni_mirror.destroy_eni_mirror_handler import DestroyEniMirrorHandler
 import aws_interactions.cloudwatch_interactions as cwi
 import aws_interactions.ec2_interactions as ec2i
-import constants as constants
+import core.constants as constants
 
 @mock.patch("lambda_destroy_eni_mirror.destroy_eni_mirror_handler.AwsClientProvider", mock.Mock())
 @mock.patch("lambda_destroy_eni_mirror.destroy_eni_mirror_handler.cwi")
